move some functions to streaming
